Module tp1::analizador::analizador_implementacion

Functions

adaptar_where 游

Esta funciom recibe un vector de strings que representa la cl치usula WHERE y devuelve un vector de expresiones.

agrupar_expresion 游

comparar_valores 游

Compara dos valores.

cumple_c_w

Eval칰a si una l칤nea cumple con la cl치usula WHERE.

evaluar_condicional 游

Eval칰a un condicional.

evaluar_expresion 游

Eval칰a una expresi칩n.

obtener_subexpresiones 游

Recibe un vector de expresiones y devuelve una expresi칩n que las agrupa.

obtener_valor 游

Obtiene el valor de la columna en la linea actual.

parsear_condicional 游

Parsea una condicional y devuelve un enum que la representa.

parsear_expresion

Parsea la cl치usula WHERE y devuelve una expresi칩n que la representa.